re structure the project code
